fix errors

+++ models.go
@@ -4,7 +4,7 @@ type Route struct {
 	BaseURL string `json:"base_url"`
 	Headers []string `json:"headers"`
 	BillingMethod string `json:"billing_method"`
-	Endpoints map[string]int32 `json:"endpoints"`
+	Endpoints map[string]int64 `json:"endpoints"`
 }
 
 type Config = map[string]Route
